feat(notifier): integrates event into checkout stepper
